refactor(server): move flyer template validation into repository
The PNG-readable, zone-shape, and zone-fits-inside-template checks belong
to the repository's invariants on the data it stores, not to the HTTP
layer's request parsing. Moving them keeps the route file focused on
multipart parsing and HTTP shape only.
