+++ server/application/src/main/kotlin/fr/devlille/partners/connect/sponsoring/application/FlyerTemplateRepositoryExposed.kt
@@ -10,9 +10,13 @@ import fr.devlille.partners.connect.sponsoring.domain.FlyerZone
 import fr.devlille.partners.connect.sponsoring.infrastructure.db.SponsoringPackEntity
 import fr.devlille.partners.connect.sponsoring.infrastructure.db.hasFlyerTemplate
 import fr.devlille.partners.connect.sponsoring.infrastructure.db.singlePackById
+import io.ktor.server.plugins.BadRequestException
 import io.ktor.server.plugins.NotFoundException
 import org.jetbrains.exposed.v1.jdbc.transactions.transaction
+import java.awt.image.BufferedImage
+import java.io.ByteArrayInputStream
 import java.util.UUID
+import javax.imageio.ImageIO
 
 class FlyerTemplateRepositoryExposed(
     private val storage: Storage,
@@ -41,19 +45,41 @@ class FlyerTemplateRepositoryExposed(
         packId: UUID,
         pngBytes: ByteArray,
         zone: FlyerZone,
-    ): FlyerTemplate = transaction {
-        val event = EventEntity.findBySlug(eventSlug)
-            ?: throw NotFoundException("Event with slug $eventSlug not found")
-        val pack = SponsoringPackEntity.singlePackById(event.id.value, packId)
-        val filename = "events/${event.id.value}/packs/$packId/flyer-template.png"
-        val upload = storage.upload(filename, pngBytes, MimeType.PNG)
-        pack.flyerTemplateUrl = upload.url
-        pack.flyerZoneX = zone.x
-        pack.flyerZoneY = zone.y
-        pack.flyerZoneWidth = zone.width
-        pack.flyerZoneHeight = zone.height
-        FlyerTemplate(templateUrl = upload.url, zone = zone)
+    ): FlyerTemplate {
+        val image = readPng(pngBytes)
+        validateZoneFitsInside(zone, image)
+        return transaction {
+            val event = EventEntity.findBySlug(eventSlug)
+                ?: throw NotFoundException("Event with slug $eventSlug not found")
+            val pack = SponsoringPackEntity.singlePackById(event.id.value, packId)
+            val filename = "events/${event.id.value}/packs/$packId/flyer-template.png"
+            val upload = storage.upload(filename, pngBytes, MimeType.PNG)
+            pack.flyerTemplateUrl = upload.url
+            pack.flyerZoneX = zone.x
+            pack.flyerZoneY = zone.y
+            pack.flyerZoneWidth = zone.width
+            pack.flyerZoneHeight = zone.height
+            FlyerTemplate(templateUrl = upload.url, zone = zone)
+        }
     }
+
+    private fun readPng(pngBytes: ByteArray): BufferedImage =
+        ImageIO.read(ByteArrayInputStream(pngBytes))
+            ?: throw BadRequestException("Uploaded file is not a readable PNG image")
+
+    private fun validateZoneFitsInside(zone: FlyerZone, image: BufferedImage) {
+        if (!zone.hasValidShape()) {
+            throw BadRequestException("Zone coordinates must be non-negative with positive width/height")
+        }
+        if (zone.x + zone.width > image.width || zone.y + zone.height > image.height) {
+            throw BadRequestException(
+                "Zone ($zone) does not fit inside template (${image.width}x${image.height})",
+            )
+        }
+    }
+
+    private fun FlyerZone.hasValidShape(): Boolean =
+        x >= 0 && y >= 0 && width > 0 && height > 0
 
     override fun clear(eventSlug: String, packId: UUID) {
         transaction {
